requirements.txt
absl-py==2.1.0
aenum==3.1.15
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
executing==2.2.0
feedparser==6.0.11
Flask==3.0.0
flatbuffers==25.2.10
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
future==1.0.0
gast==0.6.0
gnews==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
ipython==9.0.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
line-bot-sdk==3.15.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.4.1
mplfinance==0.12.10b0
multidict==6.1.0
multitasking==0.0.11
mysql-connector-python==9.2.0
namex==0.0.8
numpy==1.26.4
openai==1.63.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
prettytable==3.15.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyecharts==2.0.8
Pygments==2.19.1
pymongo==4.11.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.28.1
sentencepiece==0.2.0
setuptools==75.8.1
sgmllib3k==1.0.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
stack-data==0.6.3
ta==0.11.0
TA-Lib @ file:///C:/Users/t5566/OneDrive/Desktop/AI_Line_Bot/TA_Lib-0.4.28-cp312-cp312-win_amd64.whl#sha256=3c3db8e6661bc07aab7ed823d9d3796c9e2de0d581f109e4fcd59693fcc6e4b1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.11.1
twstock==1.4.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.1.4
yarl==1.18.3
yfinance==0.2.54
